Added: UC2_FindMaxOf3Float

+++ Day-13-Generics/MaximumNumberCheck.cs
@@ -9,7 +9,7 @@ namespace Day_13_Generics
     internal class MaximumNumberCheck
     {
 
-            public static int MaximumIntegerNumber(int firstValue, int secondValue, int thirdValue)
+            public static float MaximumIntegerNumber(float firstValue, float secondValue, float thirdValue)
             {
                 if (firstValue.CompareTo(secondValue) > 0 && firstValue.CompareTo(thirdValue) > 0)
                 {
